+++ cmd/backup.go
@@ -7,7 +7,7 @@
 package cmd
 
 import (
-	"github.com/jkaninda/mysql-bkup/pkg"
+	"github.com/jkaninda/mysql-bkup/internal"
 	"github.com/jkaninda/mysql-bkup/utils"
 	"github.com/spf13/cobra"
 )
@@ -18,7 +18,7 @@ var BackupCmd = &cobra.Command{
 	Example: utils.BackupExample,
 	Run: func(cmd *cobra.Command, args []string) {
 		if len(args) == 0 {
-			pkg.StartBackup(cmd)
+			internal.StartBackup(cmd)
 		} else {
 			utils.Fatal(`"backup" accepts no argument %q`, args)
 		}
